pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "arabic-reshaper>=3.0.0",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "mediapipe>=0.10.21",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "psycopg2-binary>=2.9.10",
    "pygame>=2.6.1",
    "python-bidi>=0.6.6",
    "pyttsx3>=2.98",
    "stripe>=12.1.0",
    "tensorflow>=2.18.1",
    "oauthlib>=3.2.2",
    "pyjwt>=2.10.1",
    "flask-dance>=7.1.0",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "blinker>=1.9.0",
    "gtts>=2.5.4",
    "requests>=2.32.3",
    "pillow>=11.2.1",
]
